requirements.txt
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.2.1
streamlit>=1.35.0
pandas>=2.0.0
